Fix issues with update from td-agent v1
Updating td-agent from v1 (fluentd v0.10.x) to v2 (fluentd v0.12.x) was
reporting as successful but not linking binaries under /usr/sbin or
adding init scripts.  As a quick and dirty fix this is just checking if
the version installed is v1 and if so it stops td-agent and removes it
before installing the new version.